# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface PropertyImage {
  id: string;
  url: string;
  alt?: string;
  isPrimary?: boolean;
}

export interface Property {
  id: string;
  title: string;
  description?: string;
  price: number;
  /** sale | rent */
  type: string;
  /** apartment | house | villa | commercial | land */
  propertyType: string;
  /** active | sold | rented | inactive */
  status: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  location: string;
  address?: string;
  city?: string;
  country?: string;
  latitude?: number | null;
  longitude?: number | null;
  amenities?: string[];
  images: PropertyImage[];
  agentName?: string | null;
  agentPhone?: string | null;
  agentEmail?: string | null;
  featured: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface PropertyListResponse {
  properties: Property[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreatePropertyInput {
  title: string;
  description?: string;
  price: number;
  type: string;
  propertyType: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  location: string;
  address?: string;
  city?: string;
  country?: string;
  latitude?: number | null;
  longitude?: number | null;
  amenities?: string[];
  imageUrls?: string[];
  agentName?: string | null;
  agentPhone?: string | null;
  agentEmail?: string | null;
  featured?: boolean;
}

export interface UpdatePropertyInput {
  title?: string;
  description?: string;
  price?: number;
  type?: string;
  propertyType?: string;
  status?: string;
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  location?: string;
  address?: string;
  city?: string;
  country?: string;
  amenities?: string[];
  imageUrls?: string[];
  agentName?: string | null;
  agentPhone?: string | null;
  agentEmail?: string | null;
  featured?: boolean;
}

export interface Inquiry {
  id: string;
  name: string;
  phone?: string | null;
  email: string;
  message: string;
  propertyId?: string | null;
  propertyTitle?: string | null;
  /** new | contacted | qualified | closed */
  status: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface InquiryListResponse {
  inquiries: Inquiry[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateInquiryInput {
  name: string;
  phone?: string | null;
  email: string;
  message: string;
  propertyId?: string | null;
}

export interface UpdateInquiryInput {
  status?: string;
  notes?: string | null;
}

export interface Visit {
  id: string;
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  propertyId?: string | null;
  propertyTitle?: string | null;
  visitDate: string;
  notes?: string | null;
  /** scheduled | completed | cancelled */
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface VisitListResponse {
  visits: Visit[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateVisitInput {
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  propertyId?: string | null;
  visitDate: string;
  notes?: string | null;
}

export interface UpdateVisitInput {
  clientName?: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  visitDate?: string;
  notes?: string | null;
  status?: string;
}

export interface Deal {
  id: string;
  propertyId?: string | null;
  propertyTitle?: string | null;
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  offerPrice: number;
  commission?: number | null;
  /** lead | negotiation | contract | closed | lost */
  dealStage: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface DealListResponse {
  deals: Deal[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateDealInput {
  propertyId?: string | null;
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  offerPrice: number;
  commission?: number | null;
  dealStage?: string;
  notes?: string | null;
}

export interface UpdateDealInput {
  propertyId?: string | null;
  clientName?: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  offerPrice?: number;
  commission?: number | null;
  dealStage?: string;
  notes?: string | null;
}

export interface DashboardAnalytics {
  totalProperties: number;
  activeProperties: number;
  soldProperties: number;
  rentedProperties: number;
  totalInquiries: number;
  activeInquiries: number;
  scheduledVisits: number;
  completedVisits: number;
  totalDeals: number;
  dealsInProgress: number;
  closedDeals: number;
  totalRevenue: number;
  recentInquiries: Inquiry[];
  recentVisits: Visit[];
}

export type ListPropertiesParams = {
  location?: string;
  type?: string;
  minPrice?: number;
  maxPrice?: number;
  bedrooms?: number;
  bathrooms?: number;
  status?: string;
  featured?: boolean;
  page?: number;
  limit?: number;
};

export type ListInquiriesParams = {
  status?: string;
  propertyId?: string;
  page?: number;
  limit?: number;
};

export type ListVisitsParams = {
  status?: string;
  propertyId?: string;
  page?: number;
  limit?: number;
};

export type ListDealsParams = {
  stage?: string;
  propertyId?: string;
  page?: number;
  limit?: number;
};