# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Real Estate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List all site visits
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitListResponse"

    post:
      operationId: createVisit
      tags: [visits]
      summary: Schedule a site visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVisitInput"
      responses:
        "201":
          description: Created visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get visit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Visit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"

    put:
      operationId: updateVisit
      tags: [visits]
      summary: Update visit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVisitInput"
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
...
components:
  schemas:
...
    Visit:
      type: object
      properties:
        id:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
          nullable: true
        clientPhone:
          type: string
          nullable: true
        propertyId:
          type: string
          nullable: true
        propertyTitle:
          type: string
          nullable: true
        visitDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        status:
          type: string
          description: "scheduled | completed | cancelled"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientName, visitDate, status, createdAt, updatedAt]
    VisitListResponse:
      type: object
      properties:
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [visits, total, page, limit, totalPages]
    CreateVisitInput:
      type: object
      properties:
        clientName:
          type: string
        clientEmail:
          type: string
          nullable: true
        clientPhone:
          type: string
          nullable: true
        propertyId:
          type: string
          nullable: true
        visitDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required: [clientName, visitDate]
    UpdateVisitInput:
      type: object
      properties:
        clientName:
          type: string
        clientEmail:
          type: string
          nullable: true
        clientPhone:
          type: string
          nullable: true
        visitDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        status:
          type: string
...