# lib/api-zod/src/generated/types/createVisitInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVisitInput {
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  propertyId?: string | null;
  visitDate: Date;
  notes?: string | null;
}