# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Real Estate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: string;
  clientName: string;
  clientEmail?: string | null;
  clientPhone?: string | null;
  propertyId?: string | null;
  propertyTitle?: string | null;
  visitDate: Date;
  notes?: string | null;
  /** scheduled | completed | cancelled */
  status: string;
  createdAt: Date;
  updatedAt: Date;
}